@@ -11,27 +11,27 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~11.2.13",
-    "@angular/common": "~11.2.13",
-    "@angular/compiler": "~11.2.13",
-    "@angular/core": "~11.2.13",
-    "@angular/forms": "~11.2.13",
-    "@angular/localize": "^11.2.13",
-    "@angular/platform-browser": "~11.2.13",
-    "@angular/platform-browser-dynamic": "~11.2.13",
-    "@angular/router": "~11.2.13",
-    "@ng-bootstrap/ng-bootstrap": "^9.1.0",
+    "@angular/animations": "^15.2.7",
+    "@angular/common": "^15.2.7",
+    "@angular/compiler": "^15.2.7",
+    "@angular/core": "^15.2.7",
+    "@angular/forms": "^15.2.7",
+    "@angular/localize": "^15.2.7",
+    "@angular/platform-browser": "^15.2.7",
+    "@angular/platform-browser-dynamic": "^15.2.7",
+    "@angular/router": "^15.2.7",
+    "@ng-bootstrap/ng-bootstrap": "^14.1.0",
     "bootstrap": "^4.5.0",
     "moment": "^2.26.0",
     "rxjs": "~6.6.0",
     "tslib": "^2.0.0",
-    "zone.js": "~0.10.2"
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.1102.12",
-    "@angular/cli": "~11.2.12",
-    "@angular/compiler-cli": "~11.2.13",
-    "@angular/language-service": "~11.2.13",
+    "@angular-devkit/build-angular": "^15.2.6",
+    "@angular/cli": "^15.2.6",
+    "@angular/compiler-cli": "^15.2.7",
+    "@angular/language-service": "^15.2.7",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^12.11.1",
@@ -46,6 +46,6 @@
     "protractor": "~7.0.0",
     "ts-node": "~7.0.0",
     "tslint": "~6.1.3",
-    "typescript": "~4.0.7"
+    "typescript": "~4.9.5"
   }
 }
