@@ -1,9 +1,12 @@
-const branch = require("./branch");
-const extract = require("./extract");
+import branch from './branch.js';
+import extract from './extract.js';
+import yargs from 'yargs';
+import { hideBin } from 'yargs/helpers';
 
 //ensure params are passed in
-const yargs = require('yargs');
-const argv = yargs
+const argv = yargs(hideBin(process.argv))
+    .scriptName("fif_api_call")
+    .usage('$0 <cmd> [args]')
     .command('branch', 'call branch api', (yargs) => {
         const argv = yargs
             .command('--dprn', '[9 digit dprn]')
@@ -15,6 +18,8 @@ const argv = yargs
     .command('extract', 'call master extract api', (yargs) => {
         extract.getMaster();
     })
+    .demandCommand(1, 'You need at least one command before moving on')
+    .help()
     .argv;
 
 
