@@ -3,43 +3,46 @@ package ca.payments.api_sample.kotlin.repository
 import ca.payments.api_sample.kotlin.config.ApiProperties
 import ca.payments.api_sample.kotlin.model.AccessTokenModel
 import ca.payments.api_sample.kotlin.model.BranchResponseModel
-import com.fasterxml.jackson.databind.ObjectMapper
-import com.fasterxml.jackson.module.kotlin.KotlinModule
+import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
+import feign.*
+import feign.jackson.JacksonDecoder
+import feign.jackson.JacksonEncoder
+import feign.okhttp.OkHttpClient
+import feign.slf4j.Slf4jLogger
 import okhttp3.Credentials
-import retrofit2.Retrofit
-import retrofit2.converter.jackson.JacksonConverterFactory
-import retrofit2.http.*
 
 private interface ApiService {
-    @FormUrlEncoded
+
     @Headers("Content-Type: application/x-www-form-urlencoded")
-    @POST(ApiProperties.apiTokenUrl)
-    suspend fun getAccessToken(@Header("Authorization") authorization: String, @Field("grant_type") body: String = "client_credentials"): AccessTokenModel
+    @RequestLine("POST ${ApiProperties.apiTokenUrl}")
+    @Body("grant_type={grantType}")
+    fun getAccessToken(@HeaderMap headers: Map<String, Any>, @Param("grantType") body: String = "client_credentials"): AccessTokenModel
 
     @Headers("Accept: application/vnd.fif.api.v1+json")
-    @GET(ApiProperties.apiBranchesUrl + "/{dprn}")
-    suspend fun getBranch(@Path("dprn") dprn: String, @Header("Authorization") authorization: String): BranchResponseModel
+    @RequestLine("GET ${ApiProperties.apiBranchesUrl}/{dprn}")
+    fun getBranch(@Param("dprn") dprn: String, @HeaderMap headers: Map<String, Any>): BranchResponseModel
 }
 
 class ApiRepository {
-    private val objectMapper by lazy {
-        val om = ObjectMapper()
-        om.registerModule(KotlinModule())
-    }
-
     private val apiService by lazy {
-        Retrofit.Builder()
-            .baseUrl(ApiProperties.apiBaseUrl)
-            .addConverterFactory(JacksonConverterFactory.create(objectMapper))
-            .build().create(ApiService::class.java)
+        val mapper = jacksonObjectMapper()
+
+        Feign.builder()
+            .client(OkHttpClient())
+            .encoder(JacksonEncoder(mapper))
+            .decoder(JacksonDecoder(mapper))
+            .logger(Slf4jLogger(this.javaClass))
+            .logLevel(Logger.Level.FULL)
+            .target(ApiService::class.java, ApiProperties.apiBaseUrl)
     }
 
     /**
      * Gets an access token using the consumer key and consumer secret
      */
-    suspend fun getAccessToken(): AccessTokenModel {
+    fun getAccessToken(): AccessTokenModel {
         val basic = Credentials.basic(ApiProperties.apiConsumerKey, ApiProperties.apiConsumerSecret)
-        return apiService.getAccessToken(basic)
+        val headers = mapOf("Authorization" to basic)
+        return apiService.getAccessToken(headers)
     }
 
     /**
@@ -47,7 +50,8 @@ class ApiRepository {
      * @param dprn: The DPRN for the branch you are looking for
      * @param accessToken: The access token retrieved from [getAccessToken]
      */
-    suspend fun getBranch(dprn: String, accessToken: String): BranchResponseModel {
-        return apiService.getBranch(dprn, "Bearer ${accessToken}")
+    fun getBranch(dprn: String, accessToken: String): BranchResponseModel {
+        val headers = mapOf("Authorization" to "Bearer ${accessToken}")
+        return apiService.getBranch(dprn, headers)
     }
 }
